requirements.txt
flask
flask_sqlalchemy
flask_bootstrap
python-dotenv
requests
python-dateutil
psycopg2-binary
